finnaly add calculation SAW

+++ src/com/main/models/bobotKriteria/insertBobotKriteria.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 import com.main.models.connectionDatabase;
 
 public class insertBobotKriteria {
-    public static boolean insertData(String kriteria, double weight, String type) {
+    public static boolean insertData(String kriteria, int weight, String type) {
         boolean data = false;
 
         String query = "INSERT INTO tbl_data_kriteria (kriteria, weight, type, lastUpdate) VALUES (?, ?, ?, NOW())";
@@ -16,7 +16,7 @@ public class insertBobotKriteria {
                 PreparedStatement state = conn.prepareStatement(query);) {
 
             state.setString(1, kriteria);
-            state.setDouble(2, weight);
+            state.setInt(2, weight);
             state.setString(3, type);
 
             if (state.executeUpdate() > 0) {
